requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==2.0.2
httpx==0.25.2
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.25.2
aioredis==2.0.1
asyncpg==0.29.0
python-multipart==0.0.6
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1